core/requirements/base.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3
drf-spectacular==0.27.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-ratelimit==4.1.0
django-cryptography==1.1
django-otp==1.2.0
django-otp[qr]==1.2.0

# Database
psycopg2-binary==2.9.9
django-geojson==3.0.0
geopy==2.4.1

# Background Tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1

# Real-time
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Email & SMS
django-anymail==13.1
twilio==8.10.0
sendgrid==6.10.0

# File Storage
boto3==1.34.0
django-storages==1.14.2

# Payment Processing
stripe==7.8.0

# AI/ML
openai==1.3.0
transformers==4.35.0
torch==2.1.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
faker==20.1.0

# Utilities
requests==2.31.0
python-decouple==3.8
Pillow==10.1.0
python-dateutil==2.8.2

# Monitoring Dependencies
psutil>=5.9.0